Fix reference error, refactor a little, moar comments.

+++ src/index.js
@@ -22,6 +22,7 @@ const NotAvailable = 'N/A';
 if (config.token) {
     // Load commands in 'commands' folder
     const commandsFolder = path.resolve(__dirname, './commands');
+    // Read all files in our commands folder that end with `.js` extension
     const commandFiles = fs.readdirSync(commandsFolder).filter(x => x.endsWith('.js'));
     for (const file of commandFiles) {
         const command = require(`./commands/${file}`);
@@ -29,6 +30,7 @@ if (config.token) {
     }
     client.on('ready', async () => {
         console.log(`Logged in as ${client.user.tag}!`);
+        // Once the bot is ready we can start updating the voice channels
         await startActiveEventsUpdater();
     });
     client.on('message', CommandHandler);
@@ -41,73 +43,69 @@ const startActiveEventsUpdater = async () => {
     if (started) return;
     setInterval(async () => {
         started = true;
-        // Get all events
-        const allEvents = await PokemonEvents.getAll();
-        // Now timestamp in seconds
-        const now = new Date() / 1000;
-        // Filter for only active evnets within todays date
-        const activeEvents = allEvents.filter(x => new Date(x.start) / 1000 < now && now < new Date(x.end) / 1000);
-        // Check if no active events available
-        if (activeEvents.length === 0) {
-            // No active events
-            return;
-        }
-        // Sort active events by end date
-        activeEvents.sort((a, b) => new Date(a.end) - new Date(b.end));
-        // Loop app specified guilds
+        // Get all active events
+        const activeEvents = await PokemonEvents.getActive(true);
+        // Loop all specified guilds
         for (const guildInfo of config.guilds) {
-            // Check if event category id set for guild
-            if (!guildInfo.eventsCategoryId) {
-                continue;
-            }
-            // Get guild from id
-            const guild = client.guilds.cache.get(guildInfo.id);
-            if (!guild) {
-                console.error(`Failed to get guild by id ${guildInfo.id}`);
-                continue;
-            }
-            // Get guild @everyone role from guild id
-            const everyoneRole = guild.roles.cache.find(x => x.id === guild.id);
-            const permissions = [{
-                id: everyoneRole.id,
-                allow: ['VIEW_CHANNEL'],
-                deny: ['CONNECT'],
-            }];
-            // Get event category channel from id
-            const channelCategory = guild.channels.cache.get(guildInfo.eventsCategoryId);
-            if (!channelCategory) {
-                console.error(`Failed to get channel category by id ${guildInfo.eventsCategoryId} from guild ${guildInfo.id}`);
-                continue;
-            }
-            // TODO: Account for ezpired event channels
-            // Update permissions for event category channel
-            await channelCategory.updateOverwrite(everyoneRole, permissions);
-            // Loop all active events
-            for (const event of activeEvents) {
-                // Format event ends date
-                const eventEndDate = event.end ? new Date(event.end) : NotAvailable;
-                // Get channel name from event name and ends date
-                const channelName = Mustache.render(config.channelNameFormat, {
-                    month: eventEndDate !== NotAvailable ? eventEndDate.getMonth() + 1 : NotAvailable,
-                    day: eventEndDate !== NotAvailable ? eventEndDate.getDate() : '',
-                    name: event.name,
-                });
-                // Check if channel exists already
-                const channelExists = guild.channels.cache.find(x => x.name.toLowerCase() === channelName.toLowerCase());
-                // Channel does not exist
-                if (!channelExists) {
-                    // Create voice channel with permissions
-                    const newChannel = await guild.channels.create(channelName, {
-                        type: 'voice',
-                        parent: channelCategory,
-                        permissionOverwrites: permissions,
-                    });
-                    console.info('Event voice channel', newChannel.name, 'created');
-                    continue;
-                }
-            }
+            createVoiceChannels(guildInfo, activeEvents);
         }
     }, intervalM);
+};
+
+const createVoiceChannels = async (guildInfo, activeEvents) => {
+    // Check if event category id set for guild
+    if (!guildInfo.eventsCategoryId) {
+        return;
+    }
+    // Get guild from id
+    const guild = client.guilds.cache.get(guildInfo.id);
+    if (!guild) {
+        console.error(`Failed to get guild by id ${guildInfo.id}`);
+        return;
+    }
+    // Get guild @everyone role from guild id
+    const everyoneRole = guild.roles.cache.find(x => x.id === guild.id);
+    const permissions = [{
+        id: everyoneRole.id,
+        allow: ['VIEW_CHANNEL'],
+        deny: ['CONNECT'],
+    }];
+    // Get event category channel from id
+    const channelCategory = guild.channels.cache.get(guildInfo.eventsCategoryId);
+    if (!channelCategory) {
+        console.error(`Failed to get channel category by id ${guildInfo.eventsCategoryId} from guild ${guildInfo.id}`);
+        return;
+    }
+    // TODO: Account for ezpired event channels
+    // Update permissions for event category channel
+    await channelCategory.updateOverwrite(everyoneRole, permissions);
+    // Loop all active events
+    for (const event of activeEvents) {
+        // Format event ends date
+        const eventEndDate = event.end ? new Date(event.end) : NotAvailable;
+        // Get channel name from event name and ends date
+        const channelName = Mustache.render(config.channelNameFormat, {
+            month: eventEndDate !== NotAvailable ? eventEndDate.getMonth() + 1 : NotAvailable,
+            day: eventEndDate !== NotAvailable ? eventEndDate.getDate() : '',
+            name: event.name,
+        });
+        createVoiceChannel(guild, channelName, channelCategory, permissions);
+    }
+};
+
+const createVoiceChannel = async (guild, channelName, channelCategory, permissions) => {
+    const channelExists = guild.channels.cache.find(x => x.name.toLowerCase() === channelName.toLowerCase());
+    // Check if channel exists already
+    if (channelExists)
+        return;
+
+    // Create voice channel with permissions
+    const newChannel = await guild.channels.create(channelName, {
+        type: 'voice',
+        parent: channelCategory,
+        permissionOverwrites: permissions,
+    });
+    console.info('Event voice channel', newChannel.name, 'created');
 };
 
 UrlWatcher(urlToWatch, intervalM, async () => {
@@ -116,17 +114,20 @@ UrlWatcher(urlToWatch, intervalM, async () => {
     // Send webhook notifications
     if (config.webhooks && config.webhooks.length > 0) {
         for (const webhook of config.webhooks) {
-            const whData = utils.getWebhookData(webhook);
-            if (whData) {
-                const guild = client.guilds.cache.get(whData.guild_id);
-                if (guild) {
-                    const channel = guild.channels.cache.get(whData.channel_id);
-                    if (channel) {
-                        await channel.bulkDelete(100);
+            // Delete previous event messages if set
+            if (config.deletePreviousEvents) {
+                const whData = utils.getWebhookData(webhook);
+                if (whData) {
+                    const guild = client.guilds.cache.get(whData.guild_id);
+                    if (guild) {
+                        const channel = guild.channels.cache.get(whData.channel_id);
+                        if (channel) {
+                            await channel.bulkDelete(100);
+                        }
                     }
                 }
+                await utils.post(webhook, payload);
             }
-            await utils.post(webhook, payload);
         }
     }
     // If bot token set we're logged into Discord bot
